M3-1434: add tags to Domains list view
