+++ src/components/Tags/index.ts
@@ -1,0 +1,1 @@
+export { default, Props as TagsProps } from './Tags';
