+++ src/features/linodes/LinodesLanding/LinodesViewWrapper.tsx
@@ -1,19 +1,15 @@
-import { splitAt } from 'ramda';
 import * as React from 'react';
 
 import { safeGetImageLabel } from 'src/utilities/safeGetImageLabel';
 
 import Paper from '@material-ui/core/Paper';
-import { StyleRulesCallback, withStyles, WithStyles } from '@material-ui/core/styles';
 import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
 import TableHead from '@material-ui/core/TableHead';
 
 import Grid from 'src/components/Grid';
-import ShowMore from 'src/components/ShowMore';
 import Table from 'src/components/Table';
 import TableRow from 'src/components/TableRow';
-import Tag from 'src/components/Tag';
 
 import { LinodeConfigSelectionDrawerCallback } from 'src/features/LinodeConfigSelectionDrawer';
 
@@ -29,78 +25,7 @@ interface Props {
     linodeId: number, linodeLabel: string) => void;
 }
 
-type ClassNames = 'root'
-  | 'tag';
-
-const styles: StyleRulesCallback<ClassNames> = (theme) => {
-  return ({
-    root: {},
-    tag: {
-      marginTop: theme.spacing.unit / 2,
-      marginRight: theme.spacing.unit,
-      padding: theme.spacing.unit / 2,
-      backgroundColor: theme.color.grey2,
-      color: theme.palette.text.primary,
-      fontFamily: 'LatoWeb',
-      '&:focus': {
-        backgroundColor: theme.color.grey2,
-      },
-      '& > span': {
-        position: 'relative',
-        top: -2,
-      },
-    },
-  });
-};
-
-type CombinedProps = Props & WithStyles<ClassNames>;
-
-class LinodesViewWrapper extends React.Component<CombinedProps, {}> {
-  renderTag = (tags: string[]) => {
-    const { classes } = this.props;
-    return tags.map(eachTag => {
-      return (
-        <Tag
-          className={classes.tag}
-          label={eachTag}
-          key={eachTag}
-          clickable={false}
-        />
-      )
-    })
-  }
-
-  renderMoreTags = (tags: string[]) => {
-    return (
-      <ShowMore
-        items={tags}
-        render={this.renderTag}
-      />
-    )
-  }
-
-  renderTagsAndMoreTags = (linodeTags: string[]) => {
-    const [visibleTags, additionalTags] = splitAt(3, linodeTags);
-    const { classes } = this.props;
-    return (
-      <React.Fragment>
-        {
-          visibleTags.map(eachTag => {
-            return (
-              <Tag
-                label={eachTag}
-                key={eachTag}
-                className={classes.tag}
-                clickable={false}
-              />
-            )
-          })
-        }
-        {!!additionalTags.length && this.renderMoreTags(additionalTags)}
-      </React.Fragment>
-    )
-  }
-
+class LinodesViewWrapper extends React.Component<Props, {}> {
   renderCards = () => {
     const { linodes, images, openConfigDrawer, toggleConfirmation } = this.props;
 
@@ -127,7 +52,6 @@ class LinodesViewWrapper extends React.Component<CombinedProps, {}> {
             linodeSpecVcpus={linode.specs.vcpus}
             linodeSpecTransfer={linode.specs.transfer}
             toggleConfirmation={toggleConfirmation}
-            renderTagsAndMoreTags={this.renderTagsAndMoreTags}
           />,
         )}
       </Grid>
@@ -168,7 +92,6 @@ class LinodesViewWrapper extends React.Component<CombinedProps, {}> {
                     linodeRecentEvent={linode.recentEvent}
                     openConfigDrawer={openConfigDrawer}
                     toggleConfirmation={toggleConfirmation}
-                    renderTagsAndMoreTags={this.renderTagsAndMoreTags}
                   />,
                 )}
               </TableBody>
@@ -191,4 +114,4 @@ class LinodesViewWrapper extends React.Component<CombinedProps, {}> {
   }
 };
 
-export default withStyles(styles, { withTheme: true })(LinodesViewWrapper);
+export default LinodesViewWrapper;
